Calculate the neighborhoods' statistics

+++ src/views/neighborhood.py
@@ -28,9 +28,12 @@ def get_all_neighborhoods():
 
 
 @neighborhood_blueprint.route(
-    '/neighborhood/<int:id_neighborhood>', methods=['DELETE'])
+    '/neighborhood/<int:id_neighborhood>', methods=['GET', 'DELETE'])
 @validate_header
-def delete_neighborhood(id_neighborhood):
-    response, status = controller.delete_neighborhood(id_neighborhood)
+def delete_get_neighborhood(id_neighborhood):
+    if request.method == 'DELETE':
+        response, status = controller.delete_neighborhood(id_neighborhood)
+    elif request.method == 'GET':
+        response, status = controller.get_one_neighborhood(id_neighborhood)
 
     return create_response(response, status)
